Edited coffee weight and exit weight in front end to receive float values and some placeholder values.

+++ Assets/Scripts/FrontEnd_Scripts/Brew List Scripts/BrewList.cs
@@ -76,7 +76,7 @@ public class BrewData
     public float coffee_weight;
     public string date;
     public int ext_time;
-    public int ext_weight;
+    public float ext_weight;
     public int grind_setting;
     public string inner_section;
     public string middle_section;
